Commit with changes

+++ src/test/java/user/UserCreateTest.java
@@ -2,6 +2,7 @@ package user;
 
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
+import org.junit.After;
 import org.junit.Test;
 import org.example.user.*;
 
@@ -10,29 +11,41 @@ public class UserCreateTest {
     private final UserClient client = new UserClient();
     private final ChecksForUser checks = new ChecksForUser();
     private final UserGenerator generator = new UserGenerator();
+    private String accessToken;
 
     @Test
     @DisplayName("Check that user can be created")
     public void userCanBeCreatedTest() {
-        Response response = client.create(generator.randomData());
-        checks.doneSuccessfully(response);
+        Response createResponse = client.create(generator.randomData());
+        checks.toDoSuccessfully(createResponse);
 
-        String accessToken = response.path("accessToken");
-        client.delete(accessToken);
+        accessToken = createResponse.path("accessToken");
     }
 
     @Test
     @DisplayName("Check that user can't be created twice with the same data")
     public void userCannotBeCreatedTwiceTest() {
-        client.create(generator.getDefault());
-        Response response = client.create(generator.getDefault());
-        checks.creatingTwiceFailed(response);
+        User user = generator.randomData();
+        client.create(user);
+        Response createResponse = client.create(user);
+        accessToken = createResponse.path("accessToken");
+        checks.toCreateTwiceFailed(createResponse);
     }
 
     @Test
     @DisplayName("Check that user can't be created without all necessary data")
     public void userCannotBeCreatedWithoutPasswordTest() {
-        Response response = client.create(generator.notFullData());
-        checks.creatingWithoutPasswordFailed(response);
+        User user = generator.randomNotFullData();
+        Response createResponse = client.create(user);
+        checks.toCreateWithoutPasswordFailed(createResponse);
+
+        accessToken = createResponse.path("accessToken");
+    }
+
+    @After
+    public void cleanUp() {
+        if (accessToken != null) {
+            client.delete(accessToken);
+        }
     }
 }
